+++ cytominer_eval/operations/replicate_reproducibility.py
@@ -12,9 +12,9 @@ from cytominer_eval.utils.transform_utils import assert_melt
 def replicate_reproducibility(
     similarity_melted_df: pd.DataFrame,
     replicate_groups: List[str],
-    quantile_over_null: np.float = 0.95,
+    quantile_over_null: float = 0.95,
     return_median_correlations: bool = False,
-) -> np.float:
+) -> float:
     r"""Summarize pairwise replicate correlations
 
     For a given pairwise similarity matrix, replicate information, and specific options,
